serverをsignal_serverに縮小しCLI一本化、READMEを全面改訂
- mars_lite/server/signal_server.py 新設: /api/signal/latest・
  /api/models系・/api/data/available・/healthのみを提供する薄いAPI。
  学習トリガー(/api/training/*)・リアルタイムメトリクス配信
  (/ws/metrics)は持たない
- mars_lite/server/metrics_server.py・training_manager.py・
  learning/training_callback.py を削除。学習はCLI
  (scripts/train_portfolio.py)に一本化
- モデル保存先の規約不一致を修正: phase_train/p0はoutput_dir直下に
  保存するが、旧metrics_serverはoutput_dir/models/を見ていたため
  server起動の学習でしか実質使えなかった。signal_serverはoutput_dir
  直下を見るよう統一し、CLI学習→signal_server配信のパスが
  実際に繋がるようにした
- scripts/run_server.py を signal_server 起動に更新
- tests/test_signal_server.py 新設: TestClientで/health・
  /api/models・/api/data/available・/api/signal/latestの200応答
  （以前のenv._recent_returns()バグの回帰テストを兼ねる）・
  prev_weights不正値の400応答・モデル無し時の404を検証
- requirements.txt: 未使用になったwebsockets依存を削除、
  TestClient用にhttpxを追加
- README.md全面改訂: CLI中心のワークフロー（fetch→ゲート→train→wf→
  serve）、シグナルサーバーのAPI一覧、RL強化実験(--obs-risk-state/
  --disagreement-dr/--phase overlay)、ディレクトリ構成を現状に合わせて
  更新。レガシー執行エージェントv1のセクションを削除
- frontend/README.md: 旧サーバー主導学習UI向けに作られたフロントエンドが
  今回のAPI縮小でどの部分が動かなくなるかを明記し、コード凍結状態である
  ことを注記（フロントエンド自体の書き直しは今回のスコープ外・今後の作業）

pytest 169 passed / 5 skipped（signal_serverのテスト10件を追加）。
p0ベンチマークはPhase 4時点との比較でビット一致（server変更は学習経路に
影響しない）。CLI学習→signal_server配信のエンドツーエンドを
TestClient+生成サンプルデータで実地確認。
